Refactoring and debugging

+++ models/user.js
@@ -64,18 +64,19 @@ userSchema
   .virtual('imageSRC')
   .get(function getImageSRC() {
     if(!this.image) return null;
+    if(this.image.match(/^http/)) return this.image;
     return `https://s3-eu-west-1.amazonaws.com/${process.env.AWS_BUCKET_NAME}/${this.image}`;
   });
 
 userSchema.pre('save', function checkPreviousImage(next) {
-  if(this.isModified('image') && this._image) {
+  if(this.isModified('image') && this._image && !this._image.match(/^http/)) {
     return s3.deleteObject({ Key: this._image }, next);
   }
   next();
 });
 
 userSchema.pre('remove', function removeImage(next) {
-  if(this.image) s3.deleteObject({ Key: this.image }, next);
+  if(this.image && !this.image.match(/^http/)) s3.deleteObject({ Key: this.image }, next);
   next();
 });
